Update timezone data file to IANA 2025b
Ingested tooling from https://github.com/fmidev/smartmet-timezones
and updated timezone data README so we have a means to regenerate
the file that is documented.

Compared to IANA 2016c, the timezone information in 2025c no longer
has a distinction between short and long names.  For example in
2016c America/Denver used MST for short, Mountain Standard Time
for long, but America/Boise used MST for both.  With 2025c they
both use MST for short and long, which also means that the built-in
"Coordinated Universal Time" long form is now displayed as "UTC"
like the short form.

While this does not fully resolve #67, it does help us maintain
the time zone data file better.